Added webrtc/audio/utility directory and empty GN target.
This Cl is the first step of a 3-step process to trick the build system of a
webrtc dependency into not failing.

The second step will be to register this folder in the dependency
build system. It cannot be done first, because there must be a GN
dependency from some existing target to a target in the new folder.

The third step is to land https://codereview.webrtc.org/2424173003, in
which the webrtc/audio/utility is used.

When a new folder with a publicly visible header is added, the build
files of said build system must be manually updated to register the
new folder.

BUG=webrtc:6548

Review-Url: https://codereview.webrtc.org/2532403003
Cr-Commit-Position: refs/heads/master@{#15299}