Enabling 'gn check' on all //webrtc/modules
We can now use a wildcard and enable 'gn check' on
all the targets in '//webrtc/modules'.

Things are changed a lot since [1] and a refactoring [2]
done by Edward has fixed the problem.

We can now close [1] and enable 'gn check' on all the
targets under '//webrtc/modules'.

[1] - https://codereview.webrtc.org/2609443003/
[2] - https://chromium.googlesource.com/external/webrtc/+/9cbb0a16e5a02538dc9cb18ec9d447e2d2058ec

BUG=webrtc:6828
NOTRY=True

Review-Url: https://codereview.webrtc.org/2732423002
Cr-Commit-Position: refs/heads/master@{#17103}