MB: Add --quick flag for low_bandwidth_audio_test
This seems to be the preferred way to pass flags to tests running on Swarming.

Add --quick flag to the low_bandwidth_audio_test and support for 'args' keyword
(similar to upstream Chromium's mb.py).
Change the type of the test to non_parallel_console_test_launcher to avoid
running with gtest-parallel, since it will swallow the stdout output.

BUG=webrtc:7229
NOTRY=True
TBR=ehmaldonado@webrtc.org

Review-Url: https://codereview.webrtc.org/2763323003
Cr-Commit-Position: refs/heads/master@{#17353}