Restore old (deprecated) signature of initializeAndroidGlobals.
This CL removed a couple parameters from the method, and changed the
type of the first parameter to an android.content.Context:
https://codereview.webrtc.org/2800353002/

But applications still using the old method may have already upcast the
context parameter to an Object, in which case this is a breaking change.

So, leaving the old signature exactly as it was before, for maximum
backwards compatibility.

BUG=webrtc:3416
TBR=magjed@webrtc.org

Review-Url: https://codereview.webrtc.org/2810973002
Cr-Commit-Position: refs/heads/master@{#17630}