Update testing tools (AppRTC, Go) to new versions
App engine is included in the AppRTC archive, so the separate archive is not needed anymore.
Mercurial is no longer necessary to include but it will be required during the vendoring step when the prebuilt archives are update.

The download/build scripts have been simplified accordingly.

The video quality test has been updated to work with the new appengine.

BUG=webrtc:7602

Review-Url: https://codereview.webrtc.org/2882073003
Cr-Commit-Position: refs/heads/master@{#18162}