Changing AudioConferenceMixer logging to base/logging.h
We'd like to remove all occurrences of WEBRTC_TRACE and delete the
macro! One logging mechanism is enough.

AudioConferenceMixer is scheduled for removal and is one of the 
things tracked by bugs.webrtc.org/4690. The logging is changed to not
block webrtc:5118

NOTRY=True

Bug: webrtc:5118
Change-Id: Ibad1ae45e8af1ba5bbe253d4c693ecf9e7c422ac
Reviewed-on: https://chromium-review.googlesource.com/518172
Commit-Queue: Alex Loiko <aleloi@webrtc.org>
Reviewed-by: Minyue Li <minyue@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#18876}