The goal of this CL is to separate Obj-C/Obj-C++ code from targets which have also C++ code (see https://bugs.chromium.org/p/webrtc/issues/detail?id=7743 for more information).
We have moved .mm files out of test_support and fileutils (into test_support_objc
and fileutils_objc).

To achieve the goal for run_test and test_renderer (in the next part of the phrase
X is run_test or test_renderer) we have created 2 targets (X_objc and X_generic)
and X will act as a proxy between these targets (this way we can avoid a circular
dependency between X_generic and X_objc).

BUG=webrtc:7743

Review-Url: https://codereview.webrtc.org/2991323003
Cr-Commit-Position: refs/heads/master@{#19479}