Keep track of the capacity delay error in the FakeNetworkPipe.
The FakeNetworkPipe use millisecond precision to calculate the delay induced by
the size of the packet being sent. The problem is that it rounds the delay down
to the closest millisecond which can cause a significant error in the actual
throughput. We keep track of that error to compensate the delay cause by
subsequent packets.

BUG=None

Review-Url: https://codereview.webrtc.org/3010653002
Cr-Commit-Position: refs/heads/master@{#19732}