Updating OpenH264 to v1.7.0
There is bug in AQ in v1.6.0 which causes raising QP to maximum value
and results in very poor quality of video no matter of allocated bitrate.
To trigger this someone needs to set packetization_mode=0 (or just do
not transmit this flag at all) in SDP. In this mode the encoder
enables multi-slice and disables AQ partially such that some part of
AQ algo still works and leads QP to maximum. The issue is fixed in v1.7.0.

BUG=webrtc:8070

Review-Url: https://codereview.webrtc.org/3011373002
Cr-Commit-Position: refs/heads/master@{#20015}