Move the TransportController from p2p/base to pc/.
The TransportController was in p2p/base before and it cannot depend on
pc/ or media/ level targets because of the circular dependency. To make the 
TransportController be responsible for creating and managing
the RtpTransport related objects which are pc/ level targets, the
TransportController is moved from p2p/base to pc/.

The TransportController makes more sense in pc/ anyway, since its main 
responsibility is processing the "transport" parts of SDP which is
PeerConnection-specific.

This is also easier than moving RtpTransport related objects to p2p/base 
because those objects also depend on other media/ and pc/ level targets
such as srtpfilter, cryptoparams etc.

Bug: webrtc:7013
Change-Id: Ic48dd5c454046ff3c81331f4b459f96a3255f328
Reviewed-on: https://webrtc-review.googlesource.com/4560
Commit-Queue: Zhi Huang <zhihuang@webrtc.org>
Reviewed-by: Taylor Brandstetter <deadbeef@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#20049}