Don't use old RTCP SR reports for remote clock estimation
At the beginning of the call, when rtt is not yet estimated, SR packets
are not used for estimation. Yet, it may happen that on some non-SR
RTCP packet RTT would become available. At that time an old SR will be
used for remote clock estimation. This will lead to remote clock offset
to the past too much.


Bug: webrtc:8468
Change-Id: I1bdbd56a7bab1c28e73987e5fb307f8e7382b045
Reviewed-on: https://webrtc-review.googlesource.com/16840
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Reviewed-by: Erik Språng <sprang@webrtc.org>
Commit-Queue: Ilya Nikolaevskiy <ilnik@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#20528}