rtc_test visibility must only be public.
"rtc_test" expands to "test" which is another template. The "test"
template overrides the visibility to public when targeting Android.

WebRTC wants to avoid this kind of situations and since a test should
be considered like an executable it is ok to force to public.

Bug: None
Change-Id: Iaf382721c60bb7345d4599a0279f94a4653247ec
No-Try: True
Reviewed-on: https://webrtc-review.googlesource.com/24560
Commit-Queue: Mirko Bonadei <mbonadei@webrtc.org>
Reviewed-by: Karl Wiberg <kwiberg@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#20813}