Test creating two senders with the same track.
When https://crbug.com/webrtc/8734 is resolved this setup should be
valid and CreateOffer() and SetLocalDescription() should work, but
currently it doesn't. It probably fails because both senders are
assigned the same ID (the track ID).

EXPECT-ing the current behavior with a TODO referencing the bug.

Bug: webrtc:8734
Change-Id: If2a9cc9b0be12c39def83b0e219e1ca82dbd7d65
Reviewed-on: https://webrtc-review.googlesource.com/39041
Reviewed-by: Steve Anton <steveanton@webrtc.org>
Commit-Queue: Henrik Boström <hbos@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#21654}