Add the rejected TransportInfo when creating an answer.
Previously, if a media section is rejected by the answerer, the TransportInfo
of that section will not be added to the answer but when answer SDP is
deserialized by the offerer, the rejected TransportInfo will be added.

This CL fixes this inconsistency by adding the TransportInfo of all the m=
sections including the rejected ones.

Bug: webrtc:8818
Change-Id: I3b163245979c4ac4df31db39262e499f1c4901c4
Reviewed-on: https://webrtc-review.googlesource.com/46380
Reviewed-by: Taylor Brandstetter <deadbeef@webrtc.org>
Commit-Queue: Zhi Huang <zhihuang@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#21815}