Removed wait from congestion window test.
Modified the StopsAndResumesMediaWhenCongestionWindowFull test so it
detects when it's test condition is fulfilled rather than waiting for a
set amount of time. This makes it less sensitive to timing changes in
the underlying code and makes it return earlier if possible.

The number of padding packets to wait for were also reduced to save some
runtime.

Bug: webrtc:8415
Change-Id: Ib72524591e976f8bea7ef8b856e2c2c87b2e13ff
Reviewed-on: https://webrtc-review.googlesource.com/56041
Commit-Queue: Sebastian Jansson <srte@webrtc.org>
Reviewed-by: Erik Språng <sprang@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#22129}