Make decoder software fallback sticky.
In practice, we never want to go back to HW decoding after deciding to
start decoding with SW codecs. This allows simplifying the code in the
fallback wrapper and makes it easier to implement HW codecs.

HW decoder is also now released when software fallback is activated
because it will not be used again. This could free up some resources.

Bug: b/73498933
Change-Id: Ibea4e32fce0c605179b649c8ac2744031799f3ee
Reviewed-on: https://webrtc-review.googlesource.com/55263
Commit-Queue: Sami Kalliomäki <sakal@webrtc.org>
Reviewed-by: Magnus Jedvert <magjed@webrtc.org>
Reviewed-by: Alex Glaznev <glaznev@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#22168}