Delete encoder caching in WebRtcVideoSendStream.
This is a followup to https://webrtc-review.googlesource.com/61640,
which ensures that picture id and tl0 pic idx are continuous,
independent of how the encoder objects are created and destroyed.

The plan is to later move responsibility for encoder creation to
VideoSendStream::ReconfigureVideoEncoder, delegating work to
VideoStreamEncoder.

Bug: webrtc:8830
Change-Id: Idde5c91f24d3c0e3fa6a3bb26eb06f6800896a28
Reviewed-on: https://webrtc-review.googlesource.com/62082
Reviewed-by: Magnus Jedvert <magjed@webrtc.org>
Commit-Queue: Niels Moller <nisse@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#22473}