Remove the limit on concurrent RtcEventLogImpl instances
The limit was introduced to avoid spawning too many threads.
It had the downside that the peer connections which have an
associated RtcEventLogImpl instance, are not necessarily those
which we wish to log.
After this CL, it becomes the responsibility of the application
hosting WebRTC to limit the number of peer connections to a
number which it can support, including thread resources.

Bug: webrtc:9046
Change-Id: I7444a6020dd51583c666285655af986def53faa4
Reviewed-on: https://webrtc-review.googlesource.com/70661
Commit-Queue: Elad Alon <eladalon@webrtc.org>
Reviewed-by: Dino Radaković <dinor@webrtc.org>
Reviewed-by: Björn Terelius <terelius@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#23019}