Update to allow the application to set a low max bitrate.
A bug surfaced when setting a low max bitrate with
30kbps hard-coded min bitrate value then a DCHECK was hit in the
VideoCodecInitializer, expecting the max bitrate to be higher than the
min bitrate. This change allows the application to set a max bitrate
below 30kbps, and adjusts the min bitrate to the value set for the
max bitrate.

RtpSender: :setParameters. If the value set was lower than the
Bug: webrtc:9141
Change-Id: I9b43ee7814b1a2caba00bc9614fc66d4438d66d8
Reviewed-on: https://webrtc-review.googlesource.com/74641
Commit-Queue: Seth Hampson <shampson@webrtc.org>
Reviewed-by: Taylor Brandstetter <deadbeef@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#23179}