Fix incorrect screenshare bitrate configs in test cases
A number of full stack tests had screenshare configured at max 2000kbps,
which differs from the 1000kbps we use in production. This CL fixes that
plus a few incorrectly set simulcast bitrates.

It also removes obsolete screenshare ALR tests (ALR is not default-on
for screenshare).

Additionally, it updates the loopback tests to not set a max cap for the
bandwidth estiamte.


Bug: chromium:861721
Change-Id: I87b9dd40f12bfc0358ccb8f8f044a6b5e51e1af5
Reviewed-on: https://webrtc-review.googlesource.com/88224
Reviewed-by: Philip Eliasson <philipel@webrtc.org>
Commit-Queue: Erik Språng <sprang@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#23935}