Fix --logs flag to webrtc_perf_tests
The was another definition of the --logs flag in full_stack_tests.cc.
The effect was that --logs set an unused bool, and the value of
FLAG_logs in test_main.cc was always false, regardless of actual
command line.

Tbr: sprang@webrtc.org
Bug: None
Change-Id: I073f8025dd897909c7e2b8d7c0ee080cb4b456ca
Reviewed-on: https://webrtc-review.googlesource.com/94900
Reviewed-by: Niels Moller <nisse@webrtc.org>
Reviewed-by: Sebastian Jansson <srte@webrtc.org>
Commit-Queue: Niels Moller <nisse@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#24347}