Add tool for aliging video files
This class adds logic for aligning a test video to a reference video
by an algorithm that maximizes SSIM between them. Aligned videos will be
easier to run video quality metrics on. This is a generic way of
aligning videos and can be replace the intrusive barcode stamping that
we currently use. This will be done in a follow-up CL.

Change-Id: I71cf1e2179c0f1e03eff9e4d8fc492fd5cfbbb1c
Bug: webrtc:9642
Reviewed-on: https://webrtc-review.googlesource.com/94773
Commit-Queue: Magnus Jedvert <magjed@webrtc.org>
Reviewed-by: Patrik Höglund <phoglund@webrtc.org>
Reviewed-by: Paulina Hensman <phensman@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#24407}