Adds fields for unacknowledged data to transport feedback.
This CL adds fields to packet feedback structs to indicate the amount of
data that was sent prior to the represented packet without being part
packet feedback, but part of bitrate allocation.

This is part of a series of CLs that allows GoogCC to track sent bitrate
that is included in bitrate allocation but without transport feedback.

Bug: webrtc:9796
Change-Id: I716a5325e2b7022ba6b3f90653542caafb056793
Reviewed-on: https://webrtc-review.googlesource.com/c/104921
Commit-Queue: Sebastian Jansson <srte@webrtc.org>
Reviewed-by: Karl Wiberg <kwiberg@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#25067}