Fix flaky JsepTransportControllerTests.
In a handful of places we wait for the old IceConnectionState to reach some value and then we assume that the new connection states have also been updated. However those are updated in response to different events that might not have fired yet, so sometimes these tests will fail.

This change makes us wait explicitly for those states to update.

Bug: webrtc:9983
Change-Id: I5cb6652ee29c0b86c0834174442140a3863e08e4
Reviewed-on: https://webrtc-review.googlesource.com/c/110441
Reviewed-by: Per Kjellander <perkj@webrtc.org>
Commit-Queue: Jonas Olsson <jonasolsson@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#25579}