Don't trigger too many probes when max allocated bitrate changes.
This fixes an issue which can happen if fec is used. The protection
rate may fluctuate and each such change would trigger a new allocation
limit to be signaled. For each such update, the probe controller could
initiate a new probe.

We work around this by both quantizing the protection fraction and by
not sending a new probe unless the max allocated bitrate has increased
significantly (or we are in ALR).

Bug: webrtc:10070
Change-Id: I328963da23aedbcbedeb877aec46f5955cd2b88d
Reviewed-on: https://webrtc-review.googlesource.com/c/113525
Reviewed-by: Sebastian Jansson <srte@webrtc.org>
Commit-Queue: Erik Språng <sprang@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#25971}