Delete test VideoSendStreamTest.CapturesTextureAndVideoFrames.
This test was added in
https://webrtc-codereview.appspot.com/15789004/, which looks like it
was early on in adding support for capture and encode with native
textures on Android. Since then, the VideoFrame/VideoFrameBuffer
interfaces have emerged and texture support has changed a lot (now
using VideoFrameBuffer with type kNative).

The test only exercises the parts of the video pipeline before
VideoStreamEncoder::pre_encode_callback_, which doesn't care at all
about the type of the corresponding VideoFrameBuffer. That's not so
useful, and since it blocks removal of pre_encode_callback, let's
delete this old test.

Bug: webrtc:9864, chromium:362437
Change-Id: I2eb6c4c48557883309fd6431bc25528441c83078
Reviewed-on: https://webrtc-review.googlesource.com/c/115411
Reviewed-by: Erik Språng <sprang@webrtc.org>
Commit-Queue: Niels Moller <nisse@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#26086}