Refactoring MID generation to use unique string generator.
This changes the MIDs that are generated if calling createOffer twice without
setting a local or remote description.
Managing the list of seen mids is now deferred to a helper object.

This is a reland of 1c376760d83119166407913b965e2e40e9d0c5f6.

> Bug: None
> Change-Id: I3440d62129884ae49aefd18e03c3a55ae096d923
> Reviewed-on: https://webrtc-review.googlesource.com/c/116021
> Commit-Queue: Amit Hilbuch <amithi@webrtc.org>
> Reviewed-by: Steve Anton <steveanton@webrtc.org>
> Reviewed-by: Seth Hampson <shampson@webrtc.org>
> Cr-Commit-Position: refs/heads/master@{#26130}

Bug: None
Change-Id: Ic8b07a252869f67a476e3af84b8072b7a130f7fd
Reviewed-on: https://webrtc-review.googlesource.com/c/116381
Reviewed-by: Steve Anton <steveanton@webrtc.org>
Commit-Queue: Amit Hilbuch <amithi@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#26151}